Reorganize files and add documentation
