+++ frontend/src/admin.js
@@ -1,3 +1,11 @@
+/**
+ * The admin page displays a list of employees with access to the database.
+ * Buttons are included to remove and add employees with access.
+ *
+ * @summary Admin page that allows users to view and edit employees with access.
+ * @author Alex Zhang
+ */
+
 import React from "react";
 import "./admin.css";
 
@@ -18,6 +26,12 @@ const entries = [
     email: "emailaddress@organization.com",
   },
 ];
+
+/**
+ * Renders the admin page
+ *
+ * @returns Contents of the admin page
+ */
 function Admin() {
   return (
     <div className="header">
@@ -26,7 +40,7 @@ function Admin() {
         {entries.map((entry) => (
           <tr key={entry.email}>
             <td className="name">
-              {entry.lastName}, {entry.firstName}
+              {entry.firstName} {entry.lastName}
             </td>
             <td className="email">{entry.email}</td>
             <button type="button">Remove</button>
